reorganise and orders tests
