+++ backend/orders/tests/test_orders.py
@@ -5,48 +5,12 @@ from rest_framework.test import APIClient
 from conftest import User, login
 from listings.models import Category, Listing
 from orders.models import Cart, CartItem
-
-
-# Fixtures
-@pytest.fixture()
-def user_fixture(request):
-    user = User.objects.create_user(
-        email="TEST@EXAMPLE.COM",
-        password="password123",
-        first_name="John",
-        last_name="Doe",
-    )
-    request.addfinalizer(user.delete)
-    return user
-
-
-@pytest.fixture()
-def listing_fixture(request, user_fixture):
-    listing = Listing.objects.create(
-        title="Test Listing",
-        image="/test.jpg",
-        description="Test Description",
-        price=100.00,
-        category=Category.objects.create(name="test", description="test"),
-        quantity=10,
-        owner_id=user_fixture.id,
-    )
-    request.addfinalizer(listing.delete)
-    return listing
-
-
-@pytest.fixture()
-def cart_fixture(user_fixture):
-    return Cart.objects.get(buyer=user_fixture)
-
-
-@pytest.fixture()
-def cart_item_fixture(request, cart_fixture, listing_fixture):
-    cart_item = CartItem.objects.create(
-        quantity=1, cart=cart_fixture, listing=listing_fixture
-    )
-    request.addfinalizer(cart_item.delete)
-    return cart_item
+from .fixtures_orders import (
+    user_fixture,
+    listing_fixture,
+    cart_fixture,
+    cart_item_fixture,
+)
 
 
 # Test the CartItem model
@@ -73,7 +37,7 @@ class TestCart:
         self, user_fixture, cart_fixture, listing_fixture, cart_item_fixture
     ):
         client = APIClient()
-        login(user_fixture, client)
+        client.force_authenticate(user_fixture)
 
         # List
         cart_item_url = reverse("cart-item-list")
@@ -90,7 +54,6 @@ class TestCart:
     @pytest.mark.django_db
     def test_add_item_to_cart(self, user_fixture, cart_fixture):
         client = APIClient()
-        login(user_fixture, client)
 
         # Create a listing from another user to add to the cart
         user = User.objects.create_user(
@@ -113,20 +76,29 @@ class TestCart:
             cart_item_url,
             data,
         )
-        assert response.status_code == 201
-        assert CartItem.objects.filter(cart=cart_fixture, listing=listing).exists()
+        assert response.status_code == 400
 
         # Try to add the same item to the cart
+        client.force_authenticate(user_fixture)
         response = client.post(
             cart_item_url,
             data,
         )
-        assert response.status_code == 403
+        assert response.status_code == 201
+        assert CartItem.objects.filter(cart=cart_fixture, listing=listing).exists()
+
+        # Try to add the same item to the cart
+        client.logout()
+        response = client.post(
+            cart_item_url,
+            data,
+        )
+        assert response.status_code == 401
 
     @pytest.mark.django_db
     def test_remove_item_from_cart(self, user_fixture, cart_fixture, listing_fixture):
         client = APIClient()
-        login(user_fixture, client)
+        client.force_authenticate(user_fixture)
 
         # Add the listing to the cart
         cart_item_url = reverse("cart-item-list")
@@ -151,7 +123,7 @@ class TestCart:
         self, user_fixture, cart_fixture, listing_fixture, cart_item_fixture
     ):
         client = APIClient()
-        login(user_fixture, client)
+        client.force_authenticate(user_fixture)
 
         # Retrieve the cart item
         cart_item = CartItem.objects.get(cart=cart_fixture, listing=listing_fixture)
